+++ requirements-dev.txt
@@ -2,3 +2,5 @@ coveralls~=2.1.2
 doctr~=1.8.0
 pdoc3~=0.9.1
 responses~=0.12.0
+tzlocal~=2.1
+iso8601~=0.1.13
